+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
-name = tool
-version = attr: tool.version.__version__
-url = https://github.com/user-cont/tool
+name = upsint
+version = attr: upsint.version.__version__
+url = https://github.com/user-cont/upsint
 description = Easy way to git.
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -28,4 +28,4 @@ packages = find:
 
 [options.entry_points]
 console_scripts =
-    tool=tool.cli:tool
+    upsint=upsint.cli:upsint
