+++ upsint/conf.py
@@ -9,7 +9,7 @@ class Conf:
     @property
     def c(self):
         if self._c is None:
-            with open(os.path.expanduser("~/.tool.json")) as fd:
+            with open(os.path.expanduser("~/.upsint.json")) as fd:
                 self._c = json.load(fd)
         return self._c
